refactor: move cli validation helpers to separate directory
